test: split recipient reminder response tests
